% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuclear_fraction_annotation.R
\name{check_bam_anno}
\alias{check_bam_anno}
\title{Check annotation and BAM file are compatible}
\usage{
check_bam_anno(bam_file, annotation)
}
\arguments{
\item{bam_file}{character, should be a character vector pointing to the BAM
file to be checked.}

\item{annotation}{character, annotation should be a character vector pointing
to the annotation file to be checked.}
}
\value{
character. The function returns a character vector with two elements.
The first element may take one of three possible values; "pass", "warning"
or "error". The second element is a message with provides extra details e.g.
if some of the sequences in the BAM file are missing from the annotation, a
message is provided that contains a sample of those sequences.
}
\description{
This function checks whether the annotation file and the BAM file
provided to \code{nuclear_fraction_annotation()} are compatible. It pulls all of
the unique values in the seqid column of the annotation and all of the
seqname values from the BAM index. It then checks to see if all of the
sequence names in the BAM file can be found in the annotation. It returns a
character vector with two elements. The first element contains either
"pass", "warning" or error ; depending on whether all, some or nonw of the
sequences were found. The second element contains a message with more
details. Note that the function was written as a simple helper function to
be called \code{dropletQC::nuclear_fraction_annotation()} and isn't intended for
more general use.
}
\keyword{internal}
