% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_damaged_cells.R
\name{identify_damaged_cells}
\alias{identify_damaged_cells}
\title{Identify damaged cells}
\usage{
identify_damaged_cells(nf_umi_ed)
}
\arguments{
\item{nf_umi_ed}{}
}
\value{

}
\description{
Identify damaged cells
}
\examples{
#1
}
