% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_damaged_cells.R
\name{plot_EM}
\alias{plot_EM}
\title{Plot EM results}
\usage{
plot_EM(em, em_c, input_cell_type, umi_thresh, nf_thresh)
}
\arguments{
\item{input_cell_type}{character, the name of the cell type}

\item{emMclust}{Mclust, result of running EM on the log10(UMI counts) and the
nuclear fraction}

\item{em_classified}{data frame, with three columns; log10(UMI counts),
nuclear fraction and a column defining each cell as a "cell" or
"damaged_cell" - this should be the output from running \code{assess_EM} on the
model results}
}
\value{
ggarrange, three ggplots combined using \code{ggpubr::ggarrange}
}
\description{
This is a helper function called intarnally by the
\code{identify_damaged_cells} function (if plots are requested). It's not
intended for general use.
}
\keyword{internal}
