% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuclear_fraction_annotation.R
\name{get_transcript_ranges}
\alias{get_transcript_ranges}
\title{Get transcript ranges}
\usage{
get_transcript_ranges(
  input_annotation,
  input_annotation_format,
  ntiles,
  verbose_output
)
}
\arguments{
\item{input_annotation}{character, should be a character vector pointing to
the annotation file}

\item{input_annotation_format}{character. Can be one of "auto", "gff3" or
"gtf". This is passed to the 'format' argument of
GenomicFeatures::makeTxDbFromGFF(). This should generally just be left as
"auto"}

\item{ntiles}{integer, the number of blocks that exons and introns should be
grouped into}

\item{verbose_output}{logical, whether to print messages}
}
\value{
GRangesList. Contains GRanges defining merged exons and introns
grouped into ~ntiles blocks. When defining intronic regions, any regions
that overlap exons are removed.  Strand is ignored when merging all
intervals.
}
\description{
This function accepts an annotation as input and returns a
GrangesList containing; GRanges defining merged exons and introns. Note that
the function was written as a simple helper function to be called
\code{dropletQC::nuclear_fraction_annotation()} and isn't intended for more
general use.
}
\keyword{internal}
