% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuclear_fraction_annotation.R
\name{get_transcript_ranges}
\alias{get_transcript_ranges}
\title{Get transcript ranges}
\usage{
get_transcript_ranges(input_annotation, input_annotation_format)
}
\arguments{
\item{input_annotation}{character, should be a character vector pointing to
the annotation file}

\item{input_annotation_format}{character. Can be one of "auto", "gff3" or
"gtf". This is passed to the 'format' argument of
GenomicFeatures::makeTxDbFromGFF(). THis should generally just be left as
"auto"}
}
\value{
list. A list is returned containing three elements;

transcript_blocks - GRanges defining intervals that contain any overlapping
genes

exons - GRangesList containing exons grouped by transcript_blocks.
Overlapping exons are merged into a single GRanges. Strand is ignored.

introns - GRangesList containing introns grouped by transcript_blocks.
Overlapping introns are merged into a single GRanges. All intronic regions
that overlap exons are removed.  Strand is ignored.
}
\description{
This function accepts an annotation as input and returns three
sets of genomic intervals;

transcript_blocks - GRanges that groups any overlapping genes into a single
GRanges

exons - GRangesList containing merged exons grouped by transcript_blocks

introns - GRangesList containing merged introns grouped by
transcript_blocks

Note that the function was written as a simple helper function to be called
\code{dropletQC::nuclear_fraction_annotation()} and isn't intended for more
general use.
}
\keyword{internal}
